re organize zerox folder structure
