+++ node-zerox/src/openAI.ts
@@ -1,6 +1,7 @@
+import axios from "axios";
+
 import { CompletionArgs, CompletionResponse } from "./types";
 import { convertKeysToSnakeCase, encodeImageToBase64 } from "./utils";
-import axios from "axios";
 
 export const getCompletion = async ({
   apiKey,
