+++ node-zerox/src/utils/common.ts
@@ -1,0 +1,56 @@
+export const camelToSnakeCase = (str: string) =>
+  str.replace(/[A-Z]/g, (letter: string) => `_${letter.toLowerCase()}`);
+
+export const convertKeysToSnakeCase = (
+  obj: Record<string, any> | null
+): Record<string, any> => {
+  if (typeof obj !== "object" || obj === null) {
+    return obj ?? {};
+  }
+
+  return Object.fromEntries(
+    Object.entries(obj).map(([key, value]) => [camelToSnakeCase(key), value])
+  );
+};
+
+export const isString = (value: string | null): value is string => {
+  return value !== null;
+};
+
+export const isValidUrl = (string: string): boolean => {
+  let url;
+  try {
+    url = new URL(string);
+  } catch (_) {
+    return false;
+  }
+  return url.protocol === "http:" || url.protocol === "https:";
+};
+
+// Strip out the ```markdown wrapper
+export const formatMarkdown = (text: string) => {
+  let formattedMarkdown = text?.trim();
+  let loopCount = 0;
+  const maxLoops = 3;
+
+  const startsWithMarkdown = formattedMarkdown.startsWith("```markdown");
+  while (startsWithMarkdown && loopCount < maxLoops) {
+    const endsWithClosing = formattedMarkdown.endsWith("```");
+
+    if (startsWithMarkdown && endsWithClosing) {
+      const outermostBlockRegex = /^```markdown\n([\s\S]*?)\n```$/;
+      const match = outermostBlockRegex.exec(formattedMarkdown);
+
+      if (match) {
+        formattedMarkdown = match[1].trim();
+        loopCount++;
+      } else {
+        break;
+      }
+    } else {
+      break;
+    }
+  }
+
+  return formattedMarkdown;
+};
