+++ node-zerox/src/utils/image.ts
@@ -1,0 +1,68 @@
+import fs from "fs-extra";
+import sharp from "sharp";
+
+interface CleanupImageProps {
+  correctOrientation: boolean;
+  imageBuffer: Buffer;
+  scheduler: Tesseract.Scheduler | null;
+  trimEdges: boolean;
+}
+
+export const encodeImageToBase64 = async (imagePath: string) => {
+  const imageBuffer = await fs.readFile(imagePath);
+  return imageBuffer.toString("base64");
+};
+
+export const cleanupImage = async ({
+  correctOrientation,
+  imageBuffer,
+  scheduler,
+  trimEdges,
+}: CleanupImageProps) => {
+  const image = sharp(imageBuffer);
+
+  // Trim extra space around the content in the image
+  if (trimEdges) {
+    image.trim();
+  }
+
+  // scheduler would always be non-null if correctOrientation is true
+  // Adding this check to satisfy typescript
+  if (correctOrientation && scheduler) {
+    const optimalRotation = await determineOptimalRotation({
+      image,
+      scheduler,
+    });
+
+    if (optimalRotation) {
+      image.rotate(optimalRotation);
+    }
+  }
+
+  // Correct the image orientation
+  const correctedBuffer = await image.toBuffer();
+  return correctedBuffer;
+};
+
+// Determine the optimal image orientation based on OCR confidence
+// Run Tesseract on 4 image orientations and compare the outputs
+const determineOptimalRotation = async ({
+  image,
+  scheduler,
+}: {
+  image: sharp.Sharp;
+  scheduler: Tesseract.Scheduler;
+}): Promise<number> => {
+  const imageBuffer = await image.toBuffer();
+  const {
+    data: { orientation_confidence, orientation_degrees },
+  } = await scheduler.addJob("detect", imageBuffer);
+
+  if (orientation_degrees) {
+    console.log(
+      `Reorienting image ${orientation_degrees} degrees (confidence: ${orientation_confidence}%)`
+    );
+    return orientation_degrees;
+  }
+  return 0;
+};
